modal pops up for each shift
