+++ client/src/components/MasterSchedule/MSRenderer.js
@@ -51,9 +51,7 @@ export default class MSRenderer extends React.Component {
               <h1>Doe</h1>
             </button>
           </div>
-          <div className="editButtonCont">
-            <EditSchedule />
-          </div>
+          <div className="editButtonCont">{/* <EditSchedule /> */}</div>
         </div>
 
         <div className="weekdayBox">
@@ -103,127 +101,123 @@ export default class MSRenderer extends React.Component {
   }
 }
 
-function Box(props) {
-  return <div>{props.text}</div>;
-}
+// function EditSchedule() {
+//   function afterOpenModal() {
+//     // references are now sync'd and can be accessed.
+//     subtitle.style.color = "#black";
+//   }
 
-function EditSchedule() {
-  function afterOpenModal() {
-    // references are now sync'd and can be accessed.
-    subtitle.style.color = "#black";
-  }
+//   function openModal() {
+//     setIsOpen(true);
+//   }
 
-  function openModal() {
-    setIsOpen(true);
-  }
+//   function closeModal() {
+//     setIsOpen(false);
+//   }
 
-  function closeModal() {
-    setIsOpen(false);
-  }
+//   const [modalIsOpen, setIsOpen] = React.useState(false);
 
-  const [modalIsOpen, setIsOpen] = React.useState(false);
+//   var subtitle;
 
-  var subtitle;
+//   const customStyles = {
+//     content: {
+//       top: "400px",
+//       left: "50%",
+//       width: "450px",
+//       height: "400px",
+//       transform: "translate(-50%, -50%)",
+//       overflow: 0,
+//     },
+//   };
 
-  const customStyles = {
-    content: {
-      top: "400px",
-      left: "50%",
-      width: "450px",
-      height: "400px",
-      transform: "translate(-50%, -50%)",
-      overflow: 0,
-    },
-  };
+//   function submitClick() {
+//     var firstName = document.getElementById("firstName");
+//     var firstNameText = firstName.value;
+//     console.log(firstNameText);
+//     var lastName = document.getElementById("lastName");
+//     var lastNameText = lastName.value;
+//     console.log(lastNameText);
+//     var email = document.getElementById("email");
+//     var emailText = email.value;
+//     console.log(emailText);
+//     fetch("http://localhost:8000/masterschedule", {
+//       method: "POST",
+//       headers: {
+//         Accept: "application/json",
+//         "Content-Type": "application/json",
+//       },
+//       body: JSON.stringify({
+//         firstName: firstNameText,
+//         lastName: lastNameText,
+//         email: emailText,
+//       }),
+//     })
+//       .then((response) => {
+//         return response.json();
+//       })
+//       .then((jsonResponse) => {
+//         console.log(jsonResponse);
+//       });
+//     function cancelClick() {
+//       console.log("doesNothingForNow");
+//     }
+//   }
 
-  function submitClick() {
-    var firstName = document.getElementById("firstName");
-    var firstNameText = firstName.value;
-    console.log(firstNameText);
-    var lastName = document.getElementById("lastName");
-    var lastNameText = lastName.value;
-    console.log(lastNameText);
-    var email = document.getElementById("email");
-    var emailText = email.value;
-    console.log(emailText);
-    fetch("http://localhost:8000/masterschedule", {
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        firstName: firstNameText,
-        lastName: lastNameText,
-        email: emailText,
-      }),
-    })
-      .then((response) => {
-        return response.json();
-      })
-      .then((jsonResponse) => {
-        console.log(jsonResponse);
-      });
-    function cancelClick() {
-      console.log("doesNothingForNow");
-    }
-  }
+//   return (
+//     <div>
+//       <button className="editButton" onClick={openModal}>
+//         Edit Schedule
+//       </button>
+//       <Modal
+//         isOpen={modalIsOpen}
+//         onAfterOpen={afterOpenModal}
+//         onRequestClose={closeModal}
+//         style={customStyles}
+//         contentLabel="Example Modal"
+//       >
+//         <div className="AllText">
+//           <h1
+//             className="AddEmpText"
+//             ref={(_subtitle) => (subtitle = _subtitle)}
+//           >
+//             Edit Master Schedule Shift
+//           </h1>
+//           <div className="shiftInfo">
+//             <div className="locationTag">
+//               <h3 className="locTag">Moffitt Floor 4</h3>
+//             </div>
+//             <div className="timeTag">
+//               <h3 className="tTag">Sunday, 12AM</h3>
+//             </div>
+//           </div>
+//           <html>
+//             <label for="cars">Select Employees:</label>
 
-  return (
-    <div>
-      <button className="editButton" onClick={openModal}>
-        Edit Schedule
-      </button>
-      <Modal
-        isOpen={modalIsOpen}
-        onAfterOpen={afterOpenModal}
-        onRequestClose={closeModal}
-        style={customStyles}
-        contentLabel="Example Modal"
-      >
-        <div className="AllText">
-          <h1
-            className="AddEmpText"
-            ref={(_subtitle) => (subtitle = _subtitle)}
-          >
-            Edit Master Schedule Shift
-          </h1>
-          <div className="shiftInfo">
-            <div className="locationTag">
-              <h3 className="locTag">Moffitt Floor 4</h3>
-            </div>
-            <div className="timeTag">
-              <h3 className="tTag">Sunday, 12AM</h3>
-            </div>
-          </div>
-          <html>
-            <label for="cars">Select Employees:</label>
-
-            <select id="cars">
-              <option value="volvo">Kathleen</option>
-              <option value="saab">Tetsu</option>
-              <option value="opel">Kat</option>
-              <option value="audi">Stephanie</option>
-            </select>
-          </html>
-        </div>
-        <div className="button-container">
-          <a href="/masterschedule">
-            <button className="CancelButton">
-              <div className="CancelHover">
-                <div className="CancelText">
-                  <h4> Cancel</h4>
-                </div>
-              </div>
-            </button>
-          </a>
-          <button className="SubmitButton" onClick={submitClick}>
-            <div className="saveText">
-              <h4>Save Changes</h4>
-            </div>
-          </button>
-        </div>
-      </Modal>
-    </div>
-  );
-}
+//             <select id="cars">
+//               <option value="volvo">Kathleen</option>
+//               <option value="saab">Tetsu</option>
+//               <option value="opel">Kat</option>
+//               <option value="audi">Stephanie</option>
+//             </select>
+//           </html>
+//         </div>
+//         <div className="button-container">
+//           <a href="/masterschedule">
+//             <button className="CancelButton">
+//               <div className="CancelHover">
+//                 <div className="CancelText">
+//                   <h4> Cancel</h4>
+//                 </div>
+//               </div>
+//             </button>
+//           </a>
+//           <button className="SubmitButton" onClick={submitClick}>
+//             <div className="saveText">
+//               <h4>Save Changes</h4>
+//             </div>
+//           </button>
+//         </div>
+//       </Modal>
+//     </div>
+//   );
+// }
